WIP: Rewrite fund setup in new architecture with saga
