+++ src/sagas/setup.js
@@ -1,0 +1,5 @@
+function* createFund();
+
+function* setup() {
+  yield takeEvery(types.CREATE, createFund);
+}
